Rename title component and props to header
