+++ packages/components/src/progressive-disclosure-panel/progressive-disclosure-panel-header/index.js
