+++ packages/components/src/progressive-disclosure-panel/style.scss
@@ -8,7 +8,7 @@
 	column-gap: $grid-unit-20;
 	row-gap: $grid-unit-30;
 
-	.components-progressive-disclosure-panel__title {
+	.components-progressive-disclosure-panel__header {
 		margin: 0;
 		display: flex;
 		align-items: center;
